+++ eoreader/products/sar/s1_product.py
@@ -95,18 +95,33 @@ class S1Product(SarProduct):
     def _set_resolution(self) -> float:
         """
         Set product default resolution (in meters)
-
-        .. WARNING:: We assume being in High Resolution (except for WV where we must be in medium resolution)
+        See here
+        <here](https://sentinel.esa.int/web/sentinel/user-guides/sentinel-1-sar/resolutions/level-1-ground-range-detected>`_
+        for more information
         """
-        root, _ = self.read_mtd()
+        def_res = None
+        # For complex data, set regular ground range resolution provided by the constructor
+        # For Sentinel data, use the best resolution available
+        if self.product_type == S1ProductType.SLC:
+            if self.sensor_mode == S1SensorMode.SM:
+                def_res = 3.5  # Full Resolution GRD
+            elif self.sensor_mode == S1SensorMode.IW:
+                def_res = 10.0  # High  Resolution GRD
+            elif self.sensor_mode == S1SensorMode.EW:
+                def_res = 25.0  # High  Resolution GRD
+            elif self.sensor_mode == S1SensorMode.WV:
+                def_res = 25.0  # Medium Resolution GRD
 
-        # Read metadata
-        try:
-            def_res = float(root.findtext(".//rangePixelSpacing"))
-        except TypeError:
-            raise InvalidProductError(
-                "rangePixelSpacing or rowSpacing not found in metadata!"
-            )
+        if not def_res:
+            root, _ = self.read_mtd()
+
+            # Read metadata
+            try:
+                def_res = float(root.findtext(".//rangePixelSpacing"))
+            except TypeError:
+                raise InvalidProductError(
+                    "rangePixelSpacing or rowSpacing not found in metadata!"
+                )
 
         return def_res